# swagger.yaml
openapi: 3.0.3
info:
  title: StudyBuddy API
  description: |
    StudyBuddy is an educational platform that provides AI-powered chat assistance, 
    quiz management, and learning analytics for students.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Admin Access
    Some endpoints require admin privileges. Admin users have elevated permissions for managing subjects, quizzes, and user feedback.
  version: 1.0.0
  contact:
    name: StudyBuddy API Support
    email: support@studybuddy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.studybuddy.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user (Admin only)
      description: Create a new user account. Requires admin privileges.
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Chat Endpoints
...
  /admin/subjects:
    get:
      tags:
        - Admin - Subjects
      summary: Get all subjects (Admin only)
      description: Retrieve all subjects in the system
      security:
        - AdminAuth: []
      responses:
        '200':
          description: Subjects retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Admin - Subjects
      summary: Create subject (Admin only)
      description: Create a new subject
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectDto'
      responses:
        '201':
          description: Subject created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/subjects/{id}:
    get:
      tags:
        - Admin - Subjects
      summary: Get subject by ID (Admin only)
      description: Retrieve a specific subject by ID
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subject ID
      responses:
        '200':
          description: Subject retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '404':
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Admin - Subjects
      summary: Update subject (Admin only)
      description: Update an existing subject
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subject ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubjectDto'
      responses:
        '200':
          description: Subject updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Admin - Subjects
      summary: Delete subject (Admin only)
      description: Delete a subject from the system
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Subject ID
      responses:
        '200':
          description: Subject deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Admin Endpoints - Topics
  /admin/topics:
    post:
      tags:
        - Admin - Topics
      summary: Add topic to subject (Admin only)
      description: Add a new topic to an existing subject
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTopicDto'
      responses:
        '201':
          description: Topic added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Admin - Topics
      summary: Update topic (Admin only)
      description: Update an existing topic
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTopicDto'
      responses:
        '200':
          description: Topic updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Admin - Topics
      summary: Delete topic (Admin only)
      description: Delete a topic from a subject
      security:
        - AdminAuth: []
      parameters:
        - name: subjectId
          in: query
          required: true
          schema:
            type: string
          description: Subject ID
        - name: topicId
          in: query
          required: true
          schema:
            type: string
          description: Topic ID
      responses:
        '200':
          description: Topic deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subject'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Admin Endpoints - Quizzes
  /admin/quizzes:
    get:
      tags:
        - Admin - Quizzes
      summary: Get all quizzes (Admin only)
      description: Retrieve all quizzes with optional filtering
      security:
        - AdminAuth: []
      parameters:
        - name: subjectId
          in: query
          required: false
          schema:
            type: string
          description: Filter by subject ID
        - name: topicId
          in: query
          required: false
          schema:
            type: string
          description: Filter by topic ID
        - name: noOfQuestions
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Number of questions to return
      responses:
        '200':
          description: Quizzes retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Quiz'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Admin - Quizzes
      summary: Create quiz (Admin only)
      description: Create a new quiz question
      security:
        - AdminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuizDto'
      responses:
        '201':
          description: Quiz created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/quizzes/{id}:
    get:
      tags:
        - Admin - Quizzes
      summary: Get quiz by ID (Admin only)
      description: Retrieve a specific quiz by ID
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Quiz ID
      responses:
        '200':
          description: Quiz retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '404':
          description: Quiz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Admin - Quizzes
      summary: Update quiz (Admin only)
      description: Update an existing quiz
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Quiz ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuizDto'
      responses:
        '200':
          description: Quiz updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quiz'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Admin - Quizzes
      summary: Delete quiz (Admin only)
      description: Delete a quiz from the system
      security:
        - AdminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Quiz ID
      responses:
        '200':
          description: Quiz deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication
  schemas:
...
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          description: User's password (minimum 8 characters)
          example: "password123"
    RegisterDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "newuser@example.com"
        password:
          type: string
          minLength: 8
          description: User's password (minimum 8 characters)
          example: "securepassword123"

    # User Models
    User:
      type: object
      properties:
        _id:
          type: string
          description: User's unique identifier
          example: "507f1f77bcf86cd799439011"
        email:
          type: string
          format: email
          description: User's email address
          example: "user@example.com"
        role:
          type: string
          description: User's role in the system
          example: "student"
        userDetails:
          type: string
          description: Reference to user details
          example: "507f1f77bcf86cd799439012"
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp

    # Chat Models
...
    TopicDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Topic name
          example: "Quadratic Equations"
        description:
          type: string
          description: Topic description
          example: "Learn about solving quadratic equations"
    SubjectDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Subject name
          example: "Mathematics"
        description:
          type: string
          description: Subject description
          example: "Mathematical concepts and problem solving"
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicDto'
          description: List of topics in the subject
    UpdateSubjectDto:
      type: object
      properties:
        name:
          type: string
          description: Subject name
        description:
          type: string
          description: Subject description
        topics:
          type: array
          items:
            $ref: '#/components/schemas/TopicDto'
          description: List of topics in the subject
    AddTopicDto:
      type: object
      required:
        - subjectId
        - topic
      properties:
        subjectId:
          type: string
          description: Subject ID to add topic to
        topic:
          $ref: '#/components/schemas/TopicDto'
    UpdateTopicDto:
      type: object
      required:
        - subjectId
        - topicId
        - topic
      properties:
        subjectId:
          type: string
          description: Subject ID
        topicId:
          type: string
          description: Topic ID to update
        topic:
          $ref: '#/components/schemas/TopicDto'
    Subject:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier
        name:
          type: string
          description: Subject name
        description:
          type: string
          description: Subject description
        topics:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                description: Topic ID
              name:
                type: string
                description: Topic name
              description:
                type: string
                description: Topic description
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Quiz DTOs and Models
    OptionDto:
      type: object
      required:
        - text
        - isCorrect
      properties:
        text:
          type: string
          description: Option text
          example: "x = 2 or x = -3"
        isCorrect:
          type: boolean
          description: Whether this option is correct
          example: true
    CreateQuizDto:
      type: object
      required:
        - question
        - options
        - subjectId
        - topicId
      properties:
        question:
          type: string
          description: Quiz question
          example: "Solve the equation x² + x - 6 = 0"
        options:
          type: array
          items:
            $ref: '#/components/schemas/OptionDto'
          description: List of answer options
        subjectId:
          type: string
          description: Subject ID this quiz belongs to
        topicId:
          type: string
          description: Topic ID this quiz belongs to
        type:
          type: string
          description: Quiz type
          example: "multiple-choice"
        difficulty:
          type: integer
          description: Difficulty level (1-5)
          example: 3
    UpdateQuizDto:
      type: object
      properties:
        question:
          type: string
          description: Quiz question
        options:
          type: array
          items:
            $ref: '#/components/schemas/OptionDto'
          description: List of answer options
        type:
          type: string
          description: Quiz type
        difficulty:
          type: integer
          description: Difficulty level (1-5)
    Quiz:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier
        question:
          type: string
          description: Quiz question
        options:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
                description: Option text
              isCorrect:
                type: boolean
                description: Whether this option is correct
        subjectId:
          type: string
          description: Subject ID
        topicId:
          type: string
          description: Topic ID
        type:
          type: string
          description: Quiz type
        difficulty:
          type: integer
          description: Difficulty level
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Error Model
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        statusCode:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error type